Modified the relational inference to better prepare for adding coequipment section
